requirements.txt
requests
tencentcloud-sdk-python
pymupdf